# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IICPC Distributed Benchmarking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  email: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  teamName?: string | null;
  totalSubmissions?: number;
  /** @nullable */
  bestScore?: number | null;
  /** @nullable */
  rank?: number | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  username?: string;
  bio?: string;
  githubUrl?: string;
  teamName?: string;
}

export type SubmissionLanguage = typeof SubmissionLanguage[keyof typeof SubmissionLanguage];


export const SubmissionLanguage = {
  cpp: 'cpp',
  rust: 'rust',
  go: 'go',
} as const;

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  building: 'building',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Submission {
  id: string;
  userId: string;
  username?: string;
  /** @nullable */
  teamName?: string | null;
  language: SubmissionLanguage;
  filename?: string;
  status: SubmissionStatus;
  /** @nullable */
  speedScore?: number | null;
  /** @nullable */
  stabilityScore?: number | null;
  /** @nullable */
  correctnessScore?: number | null;
  /** @nullable */
  compositeScore?: number | null;
  /** @nullable */
  p50Latency?: number | null;
  /** @nullable */
  p99Latency?: number | null;
  /** @nullable */
  tps?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type SubmissionInputLanguage = typeof SubmissionInputLanguage[keyof typeof SubmissionInputLanguage];


export const SubmissionInputLanguage = {
  cpp: 'cpp',
  rust: 'rust',
  go: 'go',
} as const;

export interface SubmissionInput {
  language: SubmissionInputLanguage;
  filename: string;
  code: string;
}

export interface RunConfig {
  technicalBotCount?: number;
  fundamentalBotCount?: number;
  sentimentBotCount?: number;
  durationSeconds?: number;
}

export type TestRunStatus = typeof TestRunStatus[keyof typeof TestRunStatus];


export const TestRunStatus = {
  queued: 'queued',
  building: 'building',
  running: 'running',
  scoring: 'scoring',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface TestRun {
  id: string;
  submissionId: string;
  status: TestRunStatus;
  /** @nullable */
  currentStage?: string | null;
  /** @nullable */
  progressPct?: number | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface TimeSeriesPoint {
  t: string;
  v: number;
}

export interface TelemetryData {
  submissionId: string;
  /** @nullable */
  p50Latency?: number | null;
  /** @nullable */
  p90Latency?: number | null;
  /** @nullable */
  p99Latency?: number | null;
  /** @nullable */
  peakTps?: number | null;
  /** @nullable */
  avgTps?: number | null;
  /** @nullable */
  totalOrders?: number | null;
  /** @nullable */
  fillAccuracy?: number | null;
  /** @nullable */
  uptimePct?: number | null;
  latencySeries?: TimeSeriesPoint[];
  tpsSeries?: TimeSeriesPoint[];
}

export interface LeaderboardEntry {
  rank: number;
  submissionId: string;
  userId: string;
  username: string;
  /** @nullable */
  teamName?: string | null;
  language?: string;
  compositeScore: number;
  speedScore?: number;
  stabilityScore?: number;
  correctnessScore?: number;
  /** @nullable */
  p99Latency?: number | null;
  /** @nullable */
  tps?: number | null;
  completedAt?: string;
}

export interface LeaderboardSummary {
  totalParticipants: number;
  totalSubmissions: number;
  activeRuns?: number;
  top3: LeaderboardEntry[];
  /** @nullable */
  bestP99?: number | null;
  /** @nullable */
  bestTps?: number | null;
}

export type PipelineStatusOverall = typeof PipelineStatusOverall[keyof typeof PipelineStatusOverall];


export const PipelineStatusOverall = {
  healthy: 'healthy',
  degraded: 'degraded',
  down: 'down',
} as const;

export type ComponentHealthStatus = typeof ComponentHealthStatus[keyof typeof ComponentHealthStatus];


export const ComponentHealthStatus = {
  healthy: 'healthy',
  degraded: 'degraded',
  down: 'down',
} as const;

export interface ComponentHealth {
  name: string;
  status: ComponentHealthStatus;
  description?: string;
  /** @nullable */
  latencyMs?: number | null;
  /** @nullable */
  throughput?: number | null;
}

export interface PipelineStatus {
  overall: PipelineStatusOverall;
  components: ComponentHealth[];
  activeRuns?: number;
  queueDepth?: number;
  updatedAt: string;
}

export type FunctionDefCategory = typeof FunctionDefCategory[keyof typeof FunctionDefCategory];


export const FunctionDefCategory = {
  technical: 'technical',
  fundamental: 'fundamental',
  sentiment: 'sentiment',
  orderbook: 'orderbook',
  utility: 'utility',
} as const;

export interface FunctionParam {
  name: string;
  type: string;
  description: string;
  optional?: boolean;
}

export interface FunctionDef {
  id: string;
  name: string;
  category: FunctionDefCategory;
  description: string;
  signature: string;
  codeExample?: string;
  parameters?: FunctionParam[];
  returns?: string;
  tags?: string[];
}

export type BotClassStatusClass = typeof BotClassStatusClass[keyof typeof BotClassStatusClass];


export const BotClassStatusClass = {
  technical: 'technical',
  fundamental: 'fundamental',
  sentiment: 'sentiment',
} as const;

export type BotClassStatusStatus = typeof BotClassStatusStatus[keyof typeof BotClassStatusStatus];


export const BotClassStatusStatus = {
  idle: 'idle',
  running: 'running',
  stopped: 'stopped',
} as const;

export interface BotClassStatus {
  class: BotClassStatusClass;
  count: number;
  activeCount?: number;
  status: BotClassStatusStatus;
  /** @nullable */
  ordersPerSecond?: number | null;
  description?: string;
}

export interface BotFleetStatus {
  totalBots: number;
  activeBots?: number;
  classes: BotClassStatus[];
  /** nats or kafka depending on config */
  messageTransport?: string;
  /** swarm or kubernetes depending on config */
  orchestrator?: string;
}

export type ListSubmissionsParams = {
/**
 * Filter by user
 */
userId?: string;
limit?: number;
};

export type GetLeaderboardParams = {
limit?: number;
};

export type ListPipelineRunsParams = {
limit?: number;
};

export type ListFunctionsParams = {
/**
 * Filter by category (technical, fundamental, sentiment)
 */
category?: string;
/**
 * Search query
 */
q?: string;
};