# lib/api-zod/src/generated/types/botClassStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IICPC Distributed Benchmarking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BotClassStatusClass } from './botClassStatusClass';
import type { BotClassStatusStatus } from './botClassStatusStatus';

export interface BotClassStatus {
  class: BotClassStatusClass;
  count: number;
  activeCount?: number;
  status: BotClassStatusStatus;
  /** @nullable */
  ordersPerSecond?: number | null;
  description?: string;
}