# lib/api-zod/src/generated/types/functionParam.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IICPC Distributed Benchmarking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FunctionParam {
  name: string;
  type: string;
  description: string;
  optional?: boolean;
}