# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IICPC Distributed Benchmarking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  username?: string;
  bio?: string;
  githubUrl?: string;
  teamName?: string;
}